# BookBackend/api/openAPI.yaml
openapi: "3.0.3"
info:
  title: "BookBackend API"
  description: "BookBackend API"
  version: "1.0.0"
servers:
  - url: "https://BookBackend"
paths:
...
  /books:
    get:
      tags:
        - "Books"
      summary: "GET books"
      operationId: "getAll"
      responses:
        "200":
            $ref: "#/components/responses/BookList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - "Books"
      summary: "POST books"
      operationId: "createBook"
      requestBody:
         $ref: "#/components/requestBodies/CreateBook"
      responses:
        "201":
          $ref: "#/components/responses/Book"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      security:
        - prod: []
  /books/{bookId}:
    get:
      tags:
        - "Books"
      summary: "GET books/{id}"
      operationId: "getBooksById"
      parameters:
        - $ref: "#/components/parameters/bookId"
      responses:
          "200":
            $ref: "#/components/responses/Book"
          "400":
            $ref: "#/components/responses/BadRequest"
          "401":
            $ref: "#/components/responses/Unauthorized"
    put:
        tags:
          - "Books"
        summary: "PUT books/{id}"
        operationId: "updateBook"
        parameters:
          - $ref: "#/components/parameters/bookId"
        requestBody:
            $ref: "#/components/requestBodies/UpdateUser"
        responses:
          "200":
            $ref: "#/components/responses/Book"
          "400":
            $ref: "#/components/responses/BadRequest"
          "401":
            $ref: "#/components/responses/Unauthorized"
        security:
          - prod: [ ]
    delete:
        tags:
          - "Books"
        summary: "DELETE books/{id}"
        operationId: "deleteBook"
        parameters:
            - $ref: "#/components/parameters/bookId"
        responses:
          "200":
            description: "OK"
          "401":
            $ref: "#/components/responses/Unauthorized"
        security:
          - prod: [ ]
  /books/findByAuthor:
    get:
      tags:
        - "Books"
      summary: "GET books/findByAuthor"
      operationId: "findByAuthor"
      parameters:
          - $ref: "#/components/parameters/authorId"
      responses:
        "200":
          $ref: "#/components/responses/Book"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /books/findByLanguage:
    get:
      tags:
        - "Books"
      summary: "GET books/findByLanguage"
      operationId: "findByLanguage"
      parameters:
        - name: "languageId"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          $ref: "#/components/responses/Book"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  parameters:
    bookId:
      name: bookId
      in: path
      description: ID of book to return
      required: true
      schema:
        type: integer
        format: int64
    authorId:
        name: authorId
        in: path
        description: ID of author to return
        required: true
        schema:
            type: integer
            format: int64
...
  requestBodies:
...
    UpdateUser:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UpdateUser"
    CreateBook:
      description: "Create book"
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CreateBook"
...
  responses:
...
    BookList:
      description: "OK"
      content:
        application/json:
          schema:
            type: "array"
            items:
              $ref: "#/components/schemas/Book"
...
    Book:
      description: "OK"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Book"
...
    BadRequest:
        description: "Bad Request"
        content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Error"
    Unauthorized:
        description: "Unauthorized"
        content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Error"
  schemas:
...
    Book:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1234
        language:
          type: "string"
          example: "de"
        title:
          type: "string"
          example: "The Lord of the Rings"
        isbn:
          type: "string"
          example: "978-3-16-148410-0"
        releaseDate:
          type: "string"
          format: "date-time"
        discription:
          type: "string"
          example: "Der erste Band der Trilogie"
        publisher:
          type: "string"
          example: "xyz-publisher"
        author:
          type: array
          items:
            $ref: "#/components/schemas/Author"
        label:
          type: "string"
          example: "Hörbuch"
...
    CreateBook:
      type: "object"
      properties:
        language:
          type: "string"
          example: "de"
        title:
          type: "string"
          example: "The Lord of the Rings"
        publisher:
          type: "string"
          example: "xyz-publisher"
        isbn:
          type: "string"
          example: "978-3-16-148410-0"
        discription:
          type: "string"
          example: "Der erste Band der Trilogie"
    Author:
        type: "object"
        properties:
            id:
              type: "integer"
              format: "int64"
            name:
              type: "string"
            surname:
              type: "string"
...
    UpdateUser:
        type: "object"
        properties:
          name:
            type: "string"
            example: "Max"
          surname:
            type: "string"
            example: "Mustermann"
          email:
            type: "string"
            example: "max@musterman.de"
    Error:
        type: "object"
        properties:
            message:
              type: "string"
              example: "Bad Request"
  securitySchemes:
    prod:
      scheme: bearer
      type: http
      bearerFormat: JWT